docker-env-files/app/requirements.txt
joblib==1.0.0
numpy==1.19.5
paho-mqtt==1.5.1
PyYAML==5.3.1
scipy==1.6.0
